+++ resolver/rcmd.py
@@ -1,57 +1,57 @@
 # -*- coding: utf-8 -*-
-
-from sceptre.resolvers import Resolver
-import os
 import subprocess
+from typing import Optional, Union
 
-
-def build_expression(args, profile):
-    '''
-    Build the command that will get executed. The expression always contains
-    an AWS profile. By default, the profile that the command is executed with
-    is the one that is used to run sceptre. The user can override that profile
-    by setting an alternate profile in the `profile` parameter.
-    :param args: input arguments
-    :param profile: the AWS profile to execute with
-    :return: the expression to execute
-    '''
-    if not args:
-        raise ValueError("Missing the command to execute")
-
-    if not profile:
-        profile = os.environ.get("AWS_PROFILE")
-
-    expression = args
-    if isinstance(expression, dict):
-        try:
-            expression = args['command']
-            if not expression:
-                raise ValueError("Missing the command to execute")
-        except KeyError:
-            raise KeyError("Missing the command to execute")
-
-        if 'profile' in args:
-            # override default profile
-            profile = args['profile']
-
-    expression = f"AWS_PROFILE={profile} {expression}"
-
-    return expression
+from sceptre.exceptions import SceptreException
+from sceptre.resolvers import Resolver
+from sceptre.stack import Stack
 
 
 class SceptreResolverCmd(Resolver):
+    def __init__(
+        self,
+        argument: Union[str, dict] = None,
+        stack: Stack = None,
+        *,
+        subprocess_run=subprocess.run
+    ):
+        super(SceptreResolverCmd, self).__init__(argument, stack)
+        self._subprocess_run = subprocess_run
 
-    def __init__(self, *args, **kwargs):
-        super(SceptreResolverCmd, self).__init__(*args, **kwargs)
-
-    def resolve(self):
-        '''
+    def resolve(self) -> str:
+        """
         Executes a command in an environment shell.
         :return: the resulting output from the executed command
-        '''
-        expression = build_expression(self.argument, self.stack.profile)
-        shell = os.environ.get('SHELL', '/bin/bash')
-        output = subprocess.check_output([shell, '-c', expression])
+        """
+        command, environment_variables = self._get_subprocess_envs()
+        shell = environment_variables.get('SHELL', '/bin/bash')
+        result = self._subprocess_run(
+            [shell, '-c', command],
+            env=environment_variables,
+            check=True,
+            capture_output=True
+        )
+        output = result.stdout
+
         if isinstance(output, bytes):
-            output = output.decode(os.environ.get('SHELL_ENCODING', 'utf-8'))
+            output = output.decode(environment_variables.get('SHELL_ENCODING', 'utf-8'))
+
         return output
+
+    def _get_subprocess_envs(self):
+        profile = region = role = self.stack.connection_manager.STACK_DEFAULT
+        if isinstance(self.argument, dict):
+            profile = self.argument.get('profile', profile)
+            region = self.argument.get('region', region)
+            role = self.argument.get('sceptre_role', role)
+            command: Optional[str] = self.argument.get('command', '')
+        else:
+            command: str = self.argument
+
+        if command in ('', None):
+            raise SceptreException("A command is required for the !rcmd resolver")
+
+        environment_variables = self.stack.connection_manager.create_session_environment_variables(
+            profile, region, role
+        )
+        return command, environment_variables