+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # CHANGELOG
-
 Categories: Added, Removed, Changed, Fixed, Nonfunctional, Deprecated
 
 ## Unreleased
@@ -7,6 +6,15 @@ Categories: Added, Removed, Changed, Fixed, Nonfunctional, Deprecated
 <!--- All unreleased items go here  -->
 
 <!--- Example CHANGELOG entry
+
+## 2.0.0
+
+### Added
+- Support for region and sceptre_role dict arguments
+
+### Changed
+- Updated the requirement for Sceptre to >=4.0.1
+- All AWS environment variables for the appropriate session are passed to the subprocess
 
 ## 1.2.1